Refactor StreamlineConfigurationResource and plugin custom serializer for StreamlineConfiguration
