+++ webservice/src/main/java/com/hortonworks/streamline/webservice/StreamlineApplication.java
@@ -40,6 +40,9 @@ import com.hortonworks.streamline.streams.security.authentication.StreamlineKerb
 import com.hortonworks.streamline.streams.security.impl.DefaultStreamlineAuthorizer;
 import com.hortonworks.streamline.streams.security.service.SecurityCatalogService;
 import com.hortonworks.streamline.streams.service.GenericExceptionMapper;
+
+import com.hortonworks.streamline.webservice.configurations.*;
+import com.hortonworks.streamline.webservice.resources.StreamlineConfigurationResource;
 import io.dropwizard.Application;
 import io.dropwizard.assets.AssetsBundle;
 import io.dropwizard.jetty.HttpConnectorFactory;
