+++ webservice/src/main/java/com/hortonworks/streamline/webservice/configurations/LoginConfiguration.java
@@ -1,4 +1,4 @@
-package com.hortonworks.streamline.webservice;
+package com.hortonworks.streamline.webservice.configurations;
 
 import java.util.Map;
