+++ webservice/src/main/java/com/hortonworks/streamline/webservice/configurations/ModuleConfiguration.java
@@ -13,7 +13,7 @@
   * See the License for the specific language governing permissions and
   * limitations under the License.
  **/
-package com.hortonworks.streamline.webservice;
+package com.hortonworks.streamline.webservice.configurations;
 
 import java.util.Map;
